+++ src/Page/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Helmet } from "react-helmet";
-import Header from "./Header/Header";
-import Main from "./Main/Main";
+import Header from "./Header";
+import Main from "./Main";
 
 export default () => (
   <div>
